+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Handle TYPO3 page rendering with all the usual Symfony events around as master request
+- Dispatch all the usual Symfony events around each content element as sub request
+- Handle Symfony routing before the page resolver
+- Correctly resolve locale either from `_locale` attribute or TYPO3 site language
 
 ## [2.0.1] - 2019-04-12
 ### Added